[RAevol7] Fix a lot of warnings message in the RAevol 6 and 7 code They should not change anything but some of them could have been the cause of some bugs

+++ src/libaevol/phenotype/PhenotypicTarget.h
@@ -101,7 +101,7 @@ class PhenotypicTarget
 // ==========================================================================
   //                                 Setters
   // ==========================================================================
-  void set_segmentation(int8_t nb_segments,
+  void set_segmentation(int16_t nb_segments,
                         double* boundaries,
                         PhenotypicFeature * features,
                         bool separate_segments);
